+++ resources/views/components/layouts/teach.blade.php
@@ -253,6 +253,7 @@
     </main>
 
     @livewireScriptConfig
+    @stack('scripts')
 </body>
 
 </html>